:pencil: Some work done on User destroy view

+++ resources/views/components/controllers/user/destroy.blade.php
@@ -8,6 +8,9 @@ class UserController extends Controller
 {
     public function edit(User $user)
     {
+        $role = get_role(request()->actionOn);
+        $this->authorize('accessToRole', $role);
+
         $user->delete();
         return view('actions.users.destroy');
     }
